+++ CHANGELOG.md
@@ -24,11 +24,12 @@ _the graphics engine_
 
 ---
 
-#### [0.5.6] compiler support (_00 oct 22_)
+#### [0.5.6] render graph (_00 oct 22_)
 
 - **changed** - made type names and hashes consteval
 - **changed** - switched config to consteval
 - **changed** - fixed some comments
+- **added** - first render graph generator
 
 #### [0.5.5] attachments (_17 oct 22_)